requirements.txt
streamlit
pymongo
python-dotenv
bcrypt
pypdf2
python-docx
reportlab
google-generativeai

